[FE][BOM-240] fix: 하이라이트 지워지는 문제 해결 (#165)
* Auto stash before checking out "origin/client-dev"

* refactor: 상태 명 변경
